refactor : toEntity 호출 형식으로 수정
